chore: move `tools/wpt` to `tests/wpt/runner` (#22545)
Towards #22525

---------
