+++ .dprint.json
@@ -51,8 +51,8 @@
     "third_party",
     "tools/node_compat/TODO.md",
     "tools/node_compat/node",
-    "tools/wpt/expectation.json",
-    "tools/wpt/manifest.json",
+    "tests/wpt/runner/expectation.json",
+    "tests/wpt/runner/manifest.json",
     "ext/websocket/autobahn/reports"
   ],
   "plugins": [
