+++ .github/workflows/ci.generate.ts
@@ -850,11 +850,11 @@ const ci = {
             "deno run --allow-env --allow-net --allow-read --allow-run \\",
             "        --allow-write --unstable                         \\",
             "        --lock=tools/deno.lock.json                      \\",
-            "        ./tools/wpt.ts setup",
+            "        ./tests/wpt/runner/runner.ts setup",
             "deno run --allow-env --allow-net --allow-read --allow-run \\",
             "         --allow-write --unstable                         \\",
             "         --lock=tools/deno.lock.json              \\",
-            '         ./tools/wpt.ts run --quiet --binary="$DENO_BIN"',
+            '         ./tests/wpt/runner/runner.ts run --quiet --binary="$DENO_BIN"',
           ].join("\n"),
         },
         {
@@ -867,11 +867,11 @@ const ci = {
             "deno run --allow-env --allow-net --allow-read --allow-run \\",
             "         --allow-write --unstable                         \\",
             "         --lock=tools/deno.lock.json                      \\",
-            "         ./tools/wpt.ts setup",
+            "         ./tests/wpt/runner/runner.ts setup",
             "deno run --allow-env --allow-net --allow-read --allow-run \\",
             "         --allow-write --unstable                         \\",
             "         --lock=tools/deno.lock.json                      \\",
-            "         ./tools/wpt.ts run --quiet --release             \\",
+            "         ./tests/wpt/runner/runner.ts run --quiet --release             \\",
             '                            --binary="$DENO_BIN"          \\',
             "                            --json=wpt.json               \\",
             "                            --wptreport=wptreport.json",
