+++ .github/workflows/ci.yml
@@ -519,11 +519,11 @@ jobs:
           deno run --allow-env --allow-net --allow-read --allow-run \
                   --allow-write --unstable                         \
                   --lock=tools/deno.lock.json                      \
-                  ./tools/wpt.ts setup
+                  ./tests/wpt/runner/runner.ts setup
           deno run --allow-env --allow-net --allow-read --allow-run \
                    --allow-write --unstable                         \
                    --lock=tools/deno.lock.json              \
-                   ./tools/wpt.ts run --quiet --binary="$DENO_BIN"
+                   ./tests/wpt/runner/runner.ts run --quiet --binary="$DENO_BIN"
       - name: Run web platform tests (release)
         if: '!(matrix.skip) && (matrix.wpt && matrix.profile == ''release'')'
         env:
@@ -532,11 +532,11 @@ jobs:
           deno run --allow-env --allow-net --allow-read --allow-run \
                    --allow-write --unstable                         \
                    --lock=tools/deno.lock.json                      \
-                   ./tools/wpt.ts setup
+                   ./tests/wpt/runner/runner.ts setup
           deno run --allow-env --allow-net --allow-read --allow-run \
                    --allow-write --unstable                         \
                    --lock=tools/deno.lock.json                      \
-                   ./tools/wpt.ts run --quiet --release             \
+                   ./tests/wpt/runner/runner.ts run --quiet --release             \
                                       --binary="$DENO_BIN"          \
                                       --json=wpt.json               \
                                       --wptreport=wptreport.json
