+++ .gitignore
@@ -10,7 +10,7 @@
 gclient_config.py_entries
 /target/
 /std/hash/_wasm/target
-/tools/wpt/manifest.json
+/tests/wpt/runner/manifest.json
 /third_party/
 /tests/napi/node_modules
 /tests/napi/build
@@ -25,8 +25,8 @@ gclient_config.py_entries
 /flamegraph*.svg
 
 # WPT generated cert files
-/tools/wpt/certs/index.txt*
-/tools/wpt/certs/serial*
+/tests/wpt/runner/certs/index.txt*
+/tests/wpt/runner/certs/serial*
 
 /ext/websocket/autobahn/reports
