--- tools/wpt/config.json
@@ -1,11 +1,0 @@
-{
-  "check_subdomains": false,
-  "ssl": {
-    "type": "pregenerated",
-    "pregenerated": {
-      "ca_cert_path": "../../../tools/wpt/certs/cacert.pem",
-      "host_cert_path": "../../../tools/wpt/certs/web-platform.test.pem",
-      "host_key_path": "../../../tools/wpt/certs/web-platform.test.key"
-    }
-  }
-}
